[Bug-fix]: Use prv message in absence of knowledge
When relevant knowledge can't be found, the message sent by the user
should be used to condition on for dialog generation. This bug
resulted in the dialog model conditioning on the empty knowledge string
instead, giving it no context at all for dialog generation. This
rendered incapable of conducting chit-chat, i.e, chatting without
incorporating knowledge.